Move the extracted instantiation phases to the internal package
Issue #3066. No change in behavior.

ConnectionArrayExpander and SystemOperationModeBuilder are internals of
InstantiateModel, not API, and the bundle already has an unexported
package for internals of exactly that kind. Both move there, which keeps
them out of the exported package rather than relying on Java visibility
alone, and puts the helper types the reshaping of the connection code adds
in the same place. InstantiateModel stays the entry point that carries the
compatibility promise for both.

The package documentation now covers the package as a whole; the
across-first connection traversal introduced by issue #3037 is one group
of types in it.
